Split authenticated and resolvable interfaces
